Add HomePage.style.js
